+++ src/Signup.jsx
@@ -74,10 +74,12 @@ export default function SignUp() {
         setTimeout(() => {
             setAlert(false);
         }, 5000)
+
     }, [alert, success])
 
     async function onSubmitHandler(e) {
         e.preventDefault();
+
         if (fname === "" || lname === '' || pmail === '' || branch === '' || mobile === "" || password === "" || enrollment === "" || year==="" ) {
             setAlert(true);
             setAlertmsg("Fill required details.")
@@ -113,7 +115,6 @@ export default function SignUp() {
             "personalid": pmail
         }
 
-        try {
             var response = await fetch("http://localhost:5000/signup", {
                 method: "POST",
                 headers: {
@@ -121,9 +122,15 @@ export default function SignUp() {
                     'Content-Type': 'application/json'
                 },
                 body: JSON.stringify(formData),
-            }).then(res => res.json())
+            }).then(res => res.json()).catch(()=>{
+                setAlert(true);
+                setAlertmsg("Oops! Something went wrong. please try again later.");
+            })
 
-            if (response.message === "USER_EXIST") {
+            if(!response){
+                return
+            }
+            else if (response.message === "USER_EXIST") {
                 setAlert(true);
                 setAlertmsg("User Already Exist.")
             } else if (response.message === "SUCCESS") {
@@ -136,20 +143,17 @@ export default function SignUp() {
                 setAlert(true);
                 setAlertmsg("Please try again later. or Contact Co-ordinators");
             }
-        } catch {
-            setAlert(true);
-            setAlertmsg("Oops! Something went wrong. please try again later.");
-        }
-
+    
     }
 
 
     return (
         <Container component="main" maxWidth="xs" style={{ backgroundColor: "white", borderRadius: "10px" }}>
-            <div style={{ position: 'fixed', top: "10px", zIndex: 10, width: "30vw" }} >
+            <div style={{ position: 'fixed', top: "10px", zIndex: 10, width: "25rem" }} >
                 {alert && <Alert severity="error">{alertmsg}</Alert>}
                 {success && <Alert severity="success">{alertmsg}</Alert>}
             </div>
+
             <CssBaseline />
             <div className={classes.paper}>
                 <Avatar className={classes.avatar}>
@@ -218,6 +222,7 @@ export default function SignUp() {
                                 id="Password"
                                 onChange={(e) => setPassword(e.target.value)}
                             />
+                            <p style={{color:"red"}}>  Password cannot be changed so set accordingly.</p>
                         </Grid>
                         <Grid item xs={12}>
                             <TextField
@@ -243,7 +248,7 @@ export default function SignUp() {
                             />
                         </Grid>
                         <Grid item xs={12} sm={6}>
-                            <FormControl required variant="outlined" className={classes.formControl}>
+                            <FormControl required variant="outlined" style={{width:"9rem"}} className={classes.formControl}>
                                 <InputLabel id="demo-simple-select-outlined-label">Department</InputLabel>
                                 <Select
                                     labelId="demo-simple-select-outlined-label"
@@ -265,7 +270,7 @@ export default function SignUp() {
                             </FormControl>
                         </Grid>
                        {(branch==="CO" || branch === "EL")? <Grid item xs={12} sm={6} >
-                            <FormControl required variant="outlined" className={classes.formControl}>
+                            <FormControl required variant="outlined" style={{width:"9rem"}}  className={classes.formControl}>
                                 <InputLabel id="demo-simple-select-outlined-label">Shift</InputLabel>
                                 <Select
                                     labelId="demo-simple-select-outlined-label"
@@ -281,13 +286,12 @@ export default function SignUp() {
                             </FormControl>
                         </Grid>:null}
                         <Grid item xs={12} sm={6} >
-                            <FormControl variant="outlined" className={classes.formControl}>
+                            <FormControl variant="outlined" required  style={{width:"9rem"}} className={classes.formControl}>
                                 <InputLabel id="demo-simple-select-outlined-label">Year</InputLabel>
                                 <Select
                                     labelId="demo-simple-select-outlined-label"
                                     id="demo-simple-select-outlined"
                                     value={year}
-                                    required
                                     onChange={(e) => setYear(e.target.value)}
                                     label="Year"
                                 >
@@ -301,9 +305,11 @@ export default function SignUp() {
                             <FormControlLabel
                                 control={<Checkbox required checked={checked} onChange={() => setChecked(!checked)} color="primary" />}
                                 label="I'm hereby declare that all the above details are correct and certificates will be given according to that details."
+
                             />
                         </Grid>
                     </Grid>
+
 
                     <Button
                         type="submit"
