+++ src/index.css
@@ -7,11 +7,13 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-image: url("./poster/background.png");
+  background-image: url("./poster/background.jpeg");
   background-repeat: no-repeat;
 background-size: cover;
   background-attachment: fixed;
 }
+
+
 
 .title{
   font-family: 'Abril Fatface', cursive;
@@ -203,6 +205,7 @@ code {
 	border-radius: 50%;
 	height: 135px;
   width: 135px;
+  object-fit: cover;
 	box-shadow: 4px 3px 4px 0 rgba(83, 80, 80, 0.58);
 }
